+++ commands/uninstall/__init__.py
@@ -1,7 +1,8 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
-from commands.base import CommandBase, Command, Flag
-from py_utils.emu_utils import run, warning, error, check_output, COLORS, success, input_with_options, UNINSTALL_PATH
+from commands.base import CommandBase
+from py_utils.emu_utils import run, error, input_with_options, UNINSTALL_PATH
+
 
 class Uninstall(CommandBase):
   def __init__(self):
@@ -9,7 +10,8 @@ class Uninstall(CommandBase):
     self.name = 'uninstall'
     self.description = '👋 Uninstalls emu'
 
-  def _uninstall(self):
+  @staticmethod
+  def _uninstall():
     print('Are you sure you want to uninstall emu?')
     if input_with_options(['Y', 'n'], 'n')[0] == 0:
       run(['sh', UNINSTALL_PATH])